:recycle: Code refactoring
