+++ src/store/store.js
@@ -1,13 +1,11 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 import cart from '../store/ducks/cart';
-import rootSaga from './sagas/sagas';
+import rootSaga from './sagas';
 const sagaMiddleware = createSagaMiddleware();
 
 const store = createStore(
-  combineReducers({
-    cart,
-  }),
+  combineReducers({ cart }),
   applyMiddleware(sagaMiddleware)
 );
